snack bars dialog

+++ src/app/Components/admin/admin-doctors-list/admin-doctors-list.component.ts
@@ -1,7 +1,11 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Doctor } from 'src/app/Entities/doctor';
 import { AdminService } from 'src/app/services/admin.service';
 import { CommentService } from 'src/app/services/comment.service';
+import { DialogDeleteConfirmationComponent } from '../dialog-delete-confirmation/dialog-delete-confirmation.component';
+import { MatSnackBar, MatSnackBarConfig } from '@angular/material/snack-bar';
+import { HttpErrorResponse } from '@angular/common/http';
 
 @Component({
   selector: 'app-admin-doctors-list',
@@ -9,11 +13,16 @@ import { CommentService } from 'src/app/services/comment.service';
   styleUrls: ['./admin-doctors-list.component.scss']
 })
 export class AdminDoctorsListComponent implements OnInit {
-
-  constructor(private adminService:AdminService,private commentService:CommentService) { }
+  private config: MatSnackBarConfig = new MatSnackBarConfig();
+  constructor(private dialog:MatDialog,private adminService:AdminService,private commentService:CommentService,private snackBar: MatSnackBar,
+    ) {
+    this.config.duration = 5000;
+    this.config.horizontalPosition = 'center';
+    this.config.panelClass = 'success';
+   }
 
   doctors!:Doctor[];
-
+  
   getDoctors()
   {
     this.adminService.getDoctors().subscribe(res=>{
@@ -42,10 +51,30 @@ export class AdminDoctorsListComponent implements OnInit {
 
   deleteDoctorById(id:string)
   {
-    this.adminService.deleteDoctorById(id).subscribe(res=>{
-      console.log(res);
+    const dialogRef = this.dialog.open(DialogDeleteConfirmationComponent,{width:"400px",height:"auto"});
+    dialogRef.afterClosed().subscribe(result=>{
+      if(result == true)
+      {
+        this.adminService.deleteDoctorById(id).subscribe(res=>{
+          console.log(res);
+          this.snackBar.open('Deleted Successfully','',{
+            duration:5000,
+            panelClass:'success'
+          });
+          this.getDoctors();
+        },
+        (err:HttpErrorResponse)=>{
+          console.log(err.message);
+          this.snackBar.open(err.message,'',{
+            duration:5000,
+            panelClass:'Error'
+          })
+          
+        })
+        
+      }
     })
-    this.getDoctors();
+    
   }
 
   ngOnInit(): void {
